Update Icons and Home View Widgets
Updated the home icon SVG to use white fill instead of black for better contrast against dark backgrounds. Adjusted the `TodaysTimeCounter` widget to center the text and improve readability. Refactored the `CustomDrobContainer` widget for better code readability and consistency. Modified the `HomeViewBody` to use a `ListView` for better scrolling behavior and added a new `WoroutForTheWeek` widget to the home view layout.