Add bookmark functionality to exercise elements
This commit introduces a new feature that allows users to bookmark exercises. The `ExercisesController` class has been updated to include a boolean `isSaved` property and a `changeSaved` method to toggle this property. This method also calls `update()` to notify listeners of the change.

In the `ExerciseMainElement` widget, the bookmark icon is now controlled by a `GetBuilder` that listens to the `ExercisesController`. When the user taps the bookmark icon, the `changeSaved` method is invoked, toggling the bookmark state and updating the UI to reflect whether the exercise is saved or not. The icon changes from `Icons.bookmark_border` to `Icons.bookmark` based on the `isSaved` state.