Refactor BottomNavigationBar icons for consistency and readability
Simplify the BottomNavigationBar widget by refactoring the way icons are handled within the CustomSelectedIcon class. This change removes redundant code and improves readability by consolidating the logic for selected and unselected icons. The new approach directly passes the asset path to the CustomSelectedIcon, eliminating the need for nested widget structures and enhancing maintainability. This refactor also aligns with best practices for widget composition in Flutter, making the codebase cleaner and more efficient.